resolved the spec files
