+++ src/entity/index.ts
@@ -1,8 +1,7 @@
 import { User } from './user/user';
 import { LearningObject } from './learning-object/learning-object';
 import { SubmittableLearningObject } from './learning-object/submittable-learning-object';
-import { Outcome, OutcomeSuggestion } from './outcome/outcome';
-import { Guideline } from './standard-outcome/guideline';
+import { Guideline } from './guideline/guideline';
 import { LearningOutcome } from './learning-outcome/learning-outcome';
 import { SubmittableLearningOutcome } from './learning-outcome/submittable-learning-outcome';
 import { Collection } from './collection/collection';
@@ -12,8 +11,6 @@ export {
   User,
   LearningObject,
   SubmittableLearningObject,
-  Outcome,
-  OutcomeSuggestion,
   LearningOutcome,
   SubmittableLearningOutcome,
   Guideline,
